restructured helper
