+++ src/app/[locale]/settings/_components/role-settings/RoleDialog.tsx
@@ -1,39 +1,21 @@
 import { Box, Button, Dialog, DialogActions, DialogContent, Typography } from '@mui/material';
 import { useTranslations } from 'next-intl';
-import { BaSyxRbacRule, rbacRuleActions, rbacRuleTargets } from 'lib/services/rbac-service/RbacRulesService';
+import { BaSyxRbacRule } from 'lib/services/rbac-service/RbacRulesService';
 import { useState } from 'react';
 import EditIcon from '@mui/icons-material/Edit';
 import { ArrowBack } from '@mui/icons-material';
 import { DialogCloseButton } from 'components/basics/DialogCloseButton';
 import { TargetInformationView } from 'app/[locale]/settings/_components/role-settings/target-information/TargetInformationView';
 import { deleteAndCreateRbacRule } from 'lib/services/rbac-service/RbacActions';
-import { mapTargetInformationFormModelToDto } from 'app/[locale]/settings/_components/role-settings/FormMappingHelper';
+import { mapFormModelToBaSyxRbacRule } from 'app/[locale]/settings/_components/role-settings/FormMappingHelper';
 import { useShowError } from 'lib/hooks/UseShowError';
 import { useNotificationSpawner } from 'lib/hooks/UseNotificationSpawner';
-import { RoleDialogForm } from 'app/[locale]/settings/_components/role-settings/RoleDialogForm';
+import { RoleDialogForm, RoleFormModel } from 'app/[locale]/settings/_components/role-settings/RoleDialogForm';
 
 type RoleDialogProps = {
     readonly onClose: (reload: boolean) => void;
     readonly open: boolean;
     readonly rule: BaSyxRbacRule;
-};
-
-export type ArrayOfIds = [{ id: string }];
-
-export type TargetInformationFormModel = {
-    'aas-environment': { aasIds: ArrayOfIds; submodelIds: ArrayOfIds } | undefined;
-    aas: { aasIds: ArrayOfIds } | undefined;
-    submodel: { submodelIds: ArrayOfIds; submodelElementIdShortPaths: ArrayOfIds } | undefined;
-    'concept-description': { conceptDescriptionIds: ArrayOfIds } | undefined;
-    'aas-registry': { aasIds: ArrayOfIds } | undefined;
-    'submodel-registry': { submodelIds: ArrayOfIds } | undefined;
-    'aas-discovery-service': { aasIds: ArrayOfIds; assetIds: ArrayOfIds } | undefined;
-};
-
-export type RoleFormModel = {
-    type: keyof typeof rbacRuleTargets;
-    action: (typeof rbacRuleActions)[number];
-    targetInformation: TargetInformationFormModel;
 };
 
 export const RoleDialog = (props: RoleDialogProps) => {
@@ -42,18 +24,8 @@ export const RoleDialog = (props: RoleDialogProps) => {
     const { showError } = useShowError();
     const notificationSpawner = useNotificationSpawner();
 
-    const mapFormModelToBaSyxRbacRule = (formModel: RoleFormModel): BaSyxRbacRule => {
-        const targetInformation = mapTargetInformationFormModelToDto(formModel.targetInformation, formModel.type);
-        return {
-            idShort: props.rule.idShort,
-            role: props.rule.role,
-            action: formModel.action,
-            targetInformation: targetInformation,
-        };
-    };
-
     async function onSubmit(data: RoleFormModel) {
-        const mappedDto = mapFormModelToBaSyxRbacRule(data);
+        const mappedDto = mapFormModelToBaSyxRbacRule(data, props.rule);
         const response = await deleteAndCreateRbacRule(props.rule.idShort, mappedDto);
         if (response.isSuccess) {
             notificationSpawner.spawn({
