+++ src/app/[locale]/settings/_components/role-settings/RoleDialogForm.tsx
@@ -1,13 +1,12 @@
 import { Box, Button, DialogActions, DialogContent, FormControl, MenuItem, Select, Typography } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 import { useTranslations } from 'next-intl';
-import { BaSyxRbacRule, rbacRuleActions } from 'lib/services/rbac-service/RbacRulesService';
+import { BaSyxRbacRule, rbacRuleActions, rbacRuleTargets } from 'lib/services/rbac-service/RbacRulesService';
 import { useEffect } from 'react';
 import CheckIcon from '@mui/icons-material/Check';
 import { TargetInformationForm } from 'app/[locale]/settings/_components/role-settings/target-information/TargetInformationForm';
 import { Controller, useForm } from 'react-hook-form';
-import { RoleFormModel } from 'app/[locale]/settings/_components/role-settings/RoleDialog';
-import { mapDtoToTargetInformationFormModel } from 'app/[locale]/settings/_components/role-settings/FormMappingHelper';
+import { mapBaSyxRbacRuleToFormModel } from 'app/[locale]/settings/_components/role-settings/FormMappingHelper';
 
 type RoleDialogProps = {
     readonly onSubmit: (data: RoleFormModel) => void;
@@ -15,16 +14,26 @@ type RoleDialogProps = {
     readonly rule: BaSyxRbacRule;
 };
 
+export type ArrayOfIds = [{ id: string }];
+
+export type TargetInformationFormModel = {
+    'aas-environment': { aasIds: ArrayOfIds; submodelIds: ArrayOfIds } | undefined;
+    aas: { aasIds: ArrayOfIds } | undefined;
+    submodel: { submodelIds: ArrayOfIds; submodelElementIdShortPaths: ArrayOfIds } | undefined;
+    'concept-description': { conceptDescriptionIds: ArrayOfIds } | undefined;
+    'aas-registry': { aasIds: ArrayOfIds } | undefined;
+    'submodel-registry': { submodelIds: ArrayOfIds } | undefined;
+    'aas-discovery-service': { aasIds: ArrayOfIds; assetIds: ArrayOfIds } | undefined;
+};
+
+export type RoleFormModel = {
+    type: keyof typeof rbacRuleTargets;
+    action: (typeof rbacRuleActions)[number];
+    targetInformation: TargetInformationFormModel;
+};
+
 export const RoleDialogForm = (props: RoleDialogProps) => {
     const t = useTranslations('pages.settings.roles');
-
-    const mapBaSyxRbacRuleToFormModel = (role: BaSyxRbacRule): RoleFormModel => {
-        return {
-            type: role.targetInformation['@type'],
-            action: role.action,
-            targetInformation: mapDtoToTargetInformationFormModel(role.targetInformation),
-        };
-    };
 
     const { control, handleSubmit, setValue, getValues, reset } = useForm({
         defaultValues: mapBaSyxRbacRuleToFormModel(props.rule as BaSyxRbacRule),
